docs/requirements.txt
sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinx_rtd_theme==1.3.0
sphinxcontrib-matlabdomain==0.20.2
sphinxcontrib-napoleon==0.7
sphinx-math-dollar==1.2.1
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-video==0.2.0
myst-parser==2.0.0
pyyaml==6.0.1
